requirements.txt
click>=8.1.7
Pillow>=9.3.0
tqdm>=4.64.1
tabulate>=0.9.0